requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic[email]==2.10.0
pydantic-settings==2.6.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
python-dotenv==1.0.1
httpx==0.28.0
requests==2.32.3
openpyxl==3.1.5
schedule==1.2.2

# Email notifications
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0